🐛 bgconfigのバグ修正
