+++ pages/mypage/bgconfig/index.tsx
@@ -1,5 +1,5 @@
 import type { CustomNextPage } from "next";
-import { Theme } from "src/page/mypage/theme";
+import { Theme } from "src/page/mypage/bgconfig";
 
 const ThemePage: CustomNextPage = () => <Theme />;
